requirements.txt
streamlit
numpy
opencv-python
tensorflow
Pillow
pickle5
python-dotenv
google-generativeai
langchain
langchain-google-genai
langchain-community
faiss-cpu
PyPDF2
pytesseract